# Conflicts:
#	rewrite-csharp/csharp/OpenRewrite/CSharp/Rpc/RewriteRpcServer.cs

+++ rewrite-core/src/main/java/org/openrewrite/semver/CaretRange.java
@@ -21,7 +21,7 @@ import org.openrewrite.Validated;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import static java.lang.Integer.parseInt;
+import static java.lang.Long.parseLong;
 import static java.util.Objects.requireNonNull;
 import static org.openrewrite.semver.Semver.isVersion;
 
@@ -83,24 +83,24 @@ public class CaretRange extends LatestRelease {
         }
 
         if (minor == null) {
-            upper = Integer.toString(parseInt(major) + 1);
+            upper = Long.toString(parseLong(major) + 1);
         } else if (patch == null) {
             if ("0".equals(major)) {
-                upper = "0." + (parseInt(minor) + 1);
+                upper = "0." + (parseLong(minor) + 1);
             } else {
-                upper = (parseInt(major) + 1) + ".0";
+                upper = (parseLong(major) + 1) + ".0";
             }
         } else if (micro == null) {
             if ("0".equals(major)) {
-                upper = "0." + (parseInt(minor) + 1) + ".0";
+                upper = "0." + (parseLong(minor) + 1) + ".0";
             } else {
-                upper = (parseInt(major) + 1) + ".0.0";
+                upper = (parseLong(major) + 1) + ".0.0";
             }
         } else {
             if ("0".equals(major)) {
-                upper = "0." + (parseInt(minor) + 1) + ".0.0";
+                upper = "0." + (parseLong(minor) + 1) + ".0.0";
             } else {
-                upper = (parseInt(major) + 1) + ".0.0.0";
+                upper = (parseLong(major) + 1) + ".0.0.0";
             }
         }
